Change _nano_ticks to 64-bit
Change _nano_ticks to 64-bit with the functions that operate it
for the consistency with microkernel.

Introduce *_tick_delta() functions that return 64-bit value
and *_tick_delta_32() that return low 32 bits.


Former-commit-id: 3342ddf0126a5af5a122b0c7ab4f55e5eb1eb3bc

+++ include/microkernel/k_struct.h
@@ -322,7 +322,7 @@ DECL|task|member|ktask_t task;
 DECL|task|member|ktask_t task;
 DECL|ticks|member|int32_t ticks;
 DECL|time1|member|int64_t time1;
-DECL|time2|member|int32_t time2;
+DECL|time2|member|int64_t time2;
 DECL|timer|member|K_TIMER *timer;
 DECL|timer|member|K_TIMER *timer;
 DECL|timer|member|int timer;
